+++ tests/test-expandable-dashboard-recent-comments.php
@@ -18,6 +18,8 @@ class Expandable_Dashboard_Recent_Comments_Test extends WP_UnitTestCase {
 	public function tearDown() {
 		parent::tearDown();
 
+		c2c_ExpandableDashboardRecentComments::reset();
+
 		unset( $GLOBALS['comment'] );
 
 		remove_filter( 'c2c_expandable_dashboard_recent_comments_start_expanded', '__return_true' );