+++ expandable-dashboard-recent-comments.php
@@ -2,9 +2,9 @@
 /**
  * Plugin Name: Expandable Dashboard Recent Comments
  * Version:     2.6
- * Plugin URI:  http://coffee2code.com/wp-plugins/expandable-dashboard-recent-comments/
+ * Plugin URI:  https://coffee2code.com/wp-plugins/expandable-dashboard-recent-comments/
  * Author:      Scott Reilly
- * Author URI:  http://coffee2code.com/
+ * Author URI:  https://coffee2code.com/
  * Text Domain: expandable-dashboard-recent-comments
  * License:     GPLv2 or later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html