+++ tests/test-expandable-dashboard-recent-comments.php
@@ -44,7 +44,7 @@ class Expandable_Dashboard_Recent_Comments_Test extends WP_UnitTestCase {
 	}
 
 	public function test_hooks_action_admin_menu() {
-		$this->assertEquals( 10, add_action( 'load-index.php', array( 'c2c_ExpandableDashboardRecentComments', 'do_init' ) ) );
+		$this->assertEquals( 10, has_action( 'load-index.php', array( 'c2c_ExpandableDashboardRecentComments', 'do_init' ) ) );
 	}
 
 	//