+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 * Change: Switch `is_comment_initially_expanded()` and `is_text_excerpted()` from private to protected to facilitate unit testing
 * Unit tests:
     * New: Add tests for admin dashboard comment excerpts
-    * New: Add tests for `is_comment_initially_expanded()`, `is_text_excerpted()`, and the filter `c2c_expandable_dashboard_recent_comments_start_expanded`
+    * New: Add tests for `is_comment_initially_expanded()`
+	* New: Add tests for `is_text_excerpted()`
+	* New: Add tests for `expandable_comment_excerpts()`
+	* New: Add test for the filter `c2c_expandable_dashboard_recent_comments_start_expanded`
     * Change: Remove unused test
 * Change: Note compatibility through WP 5.3+
 * Change: Update copyright date (2020)