+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## _(in-progress)_
+## 2.7 _(2020-04-30)_
 * New: Add count of comments that could be affected by either "Expand all" or "Collapse all" next to both link, respectively
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add more items to the list)
 * Change: Add more space between the expand/collapse down/up icons and their associated text